feat(integer-i64): Phase B round 10d — dedup Op*Long opcodes
Remove or rename the Op*Long opcodes that were duplicates of their
OpXxxInt counterparts post-2c:

Deleted (duplicates of the Int variant):
- OpConstLong → OpConstInt
- OpVarLong → OpVarInt
- OpPutLong → OpPutInt
- OpConvLongFromNull → OpConvIntFromNull
- OpCastIntFromLong (identity, dead)
- OpConvLongFromInt (identity, dead)
- OpCastLongFromText → OpCastIntFromText
- OpCastLongFromSingle → OpCastIntFromSingle
- OpCastLongFromFloat → OpCastIntFromFloat
- OpGetLong → OpGetInt (both dispatch to `store.get_int/get_long`
  which are byte-identical Rust code)
- OpSetLong → OpSetInt

Renamed:
- OpFormatLong → OpFormatInt
- OpFormatStackLong → OpFormatStackInt

Parser-level emission sites updated to emit the Int variant
everywhere:
- `state/codegen.rs:195` — Value::Long → OpConstInt
- `parser/operators.rs:209-224` — f#next = x path merged into the
  OpGetInt arm with the existing seek-then-set logic
- `parser/objects.rs:285,289` — f#index / f#next → OpGetInt
- `parser/collections.rs::append_data_long` — suffix "Long" → "Int"
- `generation/dispatch.rs:362` — OpFormatLong/StackLong → Int
- `compile.rs:255` — removed OpSetLong from const-literal detection

Rust-level:
- `state/text.rs::format_long / format_stack_long` methods renamed
  to `format_int / format_stack_int` (still delegate to
  `ops::format_long` which stays named for i64 width).

OPERATORS table: 245 → 234 (-11 duplicates reclaimed).  Line-number
expectations updated for the E-constant shift (369 → 349) after
11 stdlib opcode declarations were removed.

Remaining Long-suffix references (intentional):
- OpConstLongText — unrelated (it's "const-long-text" = string
  stored in a text stream), kept as-is.
- `parallel_get_long` in default/01_code.loft:907 + native.rs —
  native FFI function pair, i32 and i64 signatures still distinct.
- `FvLong` enum variant in lib/code.loft — AST-level label, kept.
- `store.get_long`/`set_long` Rust methods — internal, renamed only
  at the opcode layer.
- Stdlib Rust helpers `op_*_long` already live alongside `op_*_int`
  (OpXxxInt bodies forward to the _long fns) — kept per round-10b.

Full suite green, clippy + fmt clean.

+++ src/parser/collections.rs
@@ -600,7 +600,7 @@ use #count instead"
         state: OutputState,
     ) {
         list.push(self.cl(
-            &(start.to_owned() + "Long"),
+            &(start.to_owned() + "Int"),
             &[
                 var,
                 fmt,
